Extract ConsoleShareButton
